+++ previous-versions/assess/RELEASES.md
@@ -49,3 +49,13 @@ Based on:
 - [python v0.34.0] previous-versions/assess
 ### Releases
 - [PyPI v0.34.0] https://pypi.org/project/codat-assess/0.34.0 - previous-versions/assess
+
+## 2023-10-02 00:12:46
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.34.1] previous-versions/assess
+### Releases
+- [PyPI v0.34.1] https://pypi.org/project/codat-assess/0.34.1 - previous-versions/assess